#1: Refactor tasks
