+++ packages/next-app/pages/_app.tsx
@@ -1,8 +1,11 @@
+import {ChakraProvider, theme} from '@chakra-ui/react';
 import React from 'react';
 
 function MyApp({Component, pageProps}) {
   return (
-    <Component {...pageProps} />
+    <ChakraProvider theme={theme}>
+      <Component {...pageProps} />
+    </ChakraProvider>
   );
 }
