+++ packages/recipe-editor/src/components/Recipes/index.ts
@@ -1,1 +1,2 @@
 export * from './EditRecipeBox';
+export * from './RecipeEditor';
