+++ packages/recipe-editor/src/context/RecipeEditorServiceProvider.tsx
@@ -8,7 +8,7 @@ import {
 } from '@wagashi-backoffice/core';
 import React from 'react';
 
-interface RecipeEditorServices {
+export interface RecipeEditorServices {
   createRecipeService: CreateRecipeService;
   updateRecipeService: UpdateRecipeService;
   deleteRecipeService: DeleteRecipeService;
@@ -17,14 +17,12 @@ interface RecipeEditorServices {
   createIngredientService: CreateIngredientService;
 }
 
-const RecipeEditorServicesContext = React.createContext<undefined | RecipeEditorServices>(
-  undefined
-);
+const RecipeEditorServicesContext = React.createContext<undefined | RecipeEditorServices>(undefined);
 
 export const useRecipeEditorServices = () => {
   const values = React.useContext(RecipeEditorServicesContext);
   if (values === undefined) {
-    throw new Error('Wrap your component with an RecipeEditorServicesProvider');
+    throw new Error('Wrap your component with an RecipeEditorServiceProvider');
   }
   return values;
 };
@@ -33,17 +31,6 @@ interface IRecipeEditorServiceProvider {
   value: RecipeEditorServices;
 }
 
-export const RecipeEditorServiceProvider: React.FC<IRecipeEditorServiceProvider> = ({
-  children,
-  value,
-}) => {
-  return (
-    <RecipeEditorServicesContext.Provider
-      value={React.useMemo(() => {
-        return value;
-      }, [])}
-    >
-      {children}
-    </RecipeEditorServicesContext.Provider>
-  );
+export const RecipeEditorServiceProvider: React.FC<IRecipeEditorServiceProvider> = ({children, value}) => {
+  return <RecipeEditorServicesContext.Provider value={value}>{children}</RecipeEditorServicesContext.Provider>;
 };
